refactor main.go
* created new package pdf_creator
* Moved logic for creating pdf to pdf_creator package
